Minor changes to adhear to code standards.
